Updates on Trilinos CMAKE for trilinos integration.
1- Added compilation and test of sparse, graph, common unit tests.
2- Added compilation of perf_test executables.
   There are no tests associated with them.

+++ perf_test/sparse/spmv/matrix_market.hpp
@@ -264,7 +264,7 @@ int SparseMatrix_WriteBinaryFormat(const char* filename, OrdinalType &nrows, Ord
       if(colInd[i]>max) max = colInd[i];
     }
     if(rowPtr[row+1]>rowPtr[row]) {
-      int span = max-min;
+      size_t span = max-min;
       if(span<min_span) min_span = span;
       if(span>max_span) max_span = span;
       ave_span += span;
